:fire: Continuation of label refactor
