+++ lib/bar.js
@@ -1,34 +1,5 @@
 var Base = require('./base/default');
 var bar = module.exports = Base.extend({
-    // include missing values
-    _prepare: function () {
-        var defaults = {
-            chart: {
-                type: 'chart',
-                width: '100',
-                height: '100',
-                'font-family' : '"Open Sans", sans-serif',
-                paddingLeft: 0,
-                paddingRight: 0,
-                paddingTop: 0,
-                paddingBottom: 0
-            }
-        };
-        this._extend(defaults, this.attributes.opts);
-        this.attributes.opts = defaults;
-        return this;
-    },
-    // public function for user to set & define the graph attributes
-    attr: function (opts) {
-        opts = opts || 0;
-        // width: 200,
-        // height: 100
-        this.attributes.data = opts.data || [];
-        this.attributes.opts = opts;
-
-        return this.postRender(this._prepare()
-            ._startCycle());
-    },
     _startCycle: function () {
         var data = this.attributes.data;
         var self = this;
@@ -36,14 +7,13 @@ var bar = module.exports = Base.extend({
         var append = self.append;
         var svg;
 
-        var scale = self._defineBaseScaleProperties(data, chart);
-        paths = self._lifeCycleManager(scale, function (newScale) {
+        paths = self._lifeCycleManager(data, chart, function (newScale) {
             svg = self.make('svg',{
                 width: chart.width,
                 height: chart.height,
                 viewBox: '0 0 ' + chart.width + ' ' + chart.height,
             });
-            return self._describeBar(data, scale);
+            return self._describeBar(data, newScale);
         });
         return append(self.element,append(svg, paths));
     },
