+++ lib/base/arc.js
@@ -1,37 +1,6 @@
 var Base = require('./default');
 var arc = require('../svg/arc');
 module.exports = Base.extend({
-    // include missing values
-    _prepare: function () {
-        var self = this;
-        var defaults = {
-            chart: {
-                type: 'chart',
-                width: '100',
-                height: '100',
-                'font-family' : '"Open Sans", sans-serif',
-                paddingLeft: 0,
-                paddingRight: 0,
-                paddingTop: 0,
-                paddingBottom: 0
-            }
-        };
-        self._extend(defaults, self.attributes.opts);
-        self.attributes.opts = defaults;
-        return self;
-    },
-    // public function for user to set & define the graph attributes
-    attr: function (opts) {
-        var self = this;
-        opts = opts || 0;
-        // width: 200,
-        // height: 100
-        self.attributes.data = opts.data || [];
-        self.attributes.opts = opts;
-
-        return self.postRender(self._prepare()
-            ._startCycle());
-    },
     // parent generator that manages the svg
     _startCycle: function (){
         var self = this;
@@ -40,26 +9,29 @@ module.exports = Base.extend({
         var svg;
 
         var append = this.append;
-        // find the max width & height
-        var outerRadius = chart.outerRadius || (chart.height < chart.width ? chart.height : chart.width) / 2;
-        // converts nums to relative => total sum equals 1
-        var relativeDataSet = this._dataSetRelativeToTotal(data);
-        var scale = {
-            relativeDataSet: relativeDataSet,
-            outerRadius: outerRadius
-        };
-        self._extend(scale, chart);
-        paths = self._lifeCycleManager(scale, function (newScale) {
+        paths = self._lifeCycleManager(data, chart, function (scale) {
             svg = self.make('svg',{
                 width: chart.width,
                 height: chart.height,
                 viewBox: '0 0 ' + chart.width + ' ' + chart.height,
             });
-            return self._describePath(outerRadius, relativeDataSet, scale);
+            return self._describePath(scale.outerRadius, scale.relativeDataSet, scale);
         });
 
         return append(self.element,
                     append(svg, paths));
+    },
+    _defineBaseScaleProperties: function (data, chart) {
+        var self = this;
+        var scale = {
+            // converts nums to relative => total sum equals 1
+            relativeDataSet: self._dataSetRelativeToTotal(data),
+            // find the max width & height
+            outerRadius: chart.outerRadius || (chart.height < chart.width ? chart.height : chart.width) / 2
+        };
+
+        self._extend(scale, chart);
+        return scale;
     },
     _polarToCartesian: arc.polarToCartesian,
     _describeArc: arc.describeArc,
