+++ lib/base/common.js
@@ -1,6 +1,6 @@
 require('../utils/math');
-var utils;
 var Class = require('./class');
+var Error = require('../utils/error');
 
 var isArray = function (obj) {
     return obj instanceof Array;
@@ -11,7 +11,7 @@ var isArray = function (obj) {
  * @param  {object} object
  * @return {object} global function object 
  */
-module.exports = utils = Class.extend({
+module.exports = Class.extend({
   // default
   init: function () {
     return this;
@@ -85,13 +85,13 @@ module.exports = utils = Class.extend({
     return scale;
   },
   // TODO:: Rename lifeCycleManager, incorrect term usage
-  _lifeCycleManager: function (scale, describe) {
+  _lifeCycleManager: function (data, chart, describe) {
     var self = this;
-    var data = this.attributes.data;
     var opts = this.attributes.opts;
     var chart = opts.chart;
     var xAxis = chart.xAxis || opts.xAxis;
     var yAxis = chart.yAxis || opts.yAxis;
+    var scale = self._defineBaseScaleProperties(data, chart);
 
     if ((chart.type != 'bubble-point') && (yAxis || xAxis)) {
       self._getLabelConfig(scale, yAxis, xAxis);
