+++ lib/base/default.js
@@ -21,5 +21,36 @@ var base = module.exports = Common.extend({
       self.token = self._makeToken();
       self.attributes = {};
       return self;
+    },
+    // include missing values
+    _prepare: function () {
+        var self = this;
+        var defaults = {
+            chart: {
+                type: 'chart',
+                width: '100',
+                height: '100',
+                'font-family' : '"Open Sans", sans-serif',
+                paddingLeft: 0,
+                paddingRight: 0,
+                paddingTop: 0,
+                paddingBottom: 0
+            }
+        };
+        self._extend(defaults, self.attributes.opts);
+        self.attributes.opts = defaults;
+        return self;
+    },
+    // public function for user to set & define the graph attributes
+    attr: function (opts) {
+        var self = this;
+        opts = opts || 0;
+        // width: 200,
+        // height: 100
+        self.attributes.data = opts.data || [];
+        self.attributes.opts = opts;
+
+        return self.postRender(self._prepare()
+            ._startCycle());
     }
 });
