+++ lib/bubble.js
@@ -1,4 +1,4 @@
-var Base = require('./base/arc');
+var Base = require('./base/default');
 var bubble = module.exports = Base.extend({
     _startCycle: function () {
         var self = this;
@@ -20,16 +20,15 @@ var bubble = module.exports = Base.extend({
 
         if (chart.type == 'scattered') {
             chart.type = 'bubble-scattered';
-            scale = self._defineBaseScaleProperties(data, chart);
-            paths = self._lifeCycleManager(scale, function (newScale) {
+            paths = self._lifeCycleManager(data, chart, function (newScale) {
                 svg = getSvg();
                 return self._describeBubbleChart(data, newScale);
             });
             return append(self.element,append(svg, paths));
         } else {
             chart.type = 'bubble-point';
-            scale = self._defineBaseScaleProperties(data, chart);
-            paths = self._lifeCycleManager(scale, function (newScale) {
+            // console.log(chart);
+            paths = self._lifeCycleManager(data, chart, function (newScale) {
                 svg = getSvg();
                 paths = self._describeBubble(data, chart.height, chart.width, newScale);
                 paths.unshift(self._describeXAxis(chart.height, chart.width, newScale));
