+++ lib/spark.js
@@ -59,22 +59,15 @@ var spark = module.exports = Base.extend({
       data = [data];
     }
 
-    var scale = self._scale(data, opts);
-    self._extend(scale, chart);
-
-    if (yAxis || xAxis) {
-      self._getLabelConfig(scale, yAxis, xAxis);
-      if (!self.describeYAxis) {
-        Error.label();
-      }
+    if (xAxis) {
+      chart.xAxis = xAxis;
     }
 
-    scale.pHeight = scale.height - scale.paddingTop - scale.paddingBottom;
-    scale.pWidth = scale.width - scale.paddingLeft - scale.paddingRight;
-    scale.heightRatio = scale.pHeight / scale.max;
-    scale.tickSize = self._sigFigs((scale.pWidth / (scale.len - 1)),8);
+    if (yAxis) {
+      chart.yAxis = yAxis;
+    }
 
-    paths = self._lifeCycleManager(scale, function (newScale) {
+    paths = self._lifeCycleManager(data, chart, function (scale) {
         svg = self.make('svg',{
             width: chart.width,
             height: chart.height,
@@ -93,6 +86,14 @@ var spark = module.exports = Base.extend({
         return paths;
     });
     return append(self.element,append(svg, paths));
+  },
+  _getRatio: function (scale) {
+    var self = this;
+
+    scale.pHeight = scale.height - scale.paddingTop - scale.paddingBottom;
+    scale.pWidth = scale.width - scale.paddingLeft - scale.paddingRight;
+    scale.heightRatio = scale.pHeight / scale.max;
+    scale.tickSize = self._sigFigs((scale.pWidth / (scale.len - 1)),8);
   },
   // describes an open path
   _describeAttributeD: function (numArr, paddingLeft, paddingTop, scale) {
